+++ app/src/main/java/org/jellyfin/androidtv/ui/settings/screen/playback/SettingsPlaybackCodecScreen.kt
@@ -24,8 +24,8 @@ fun SettingsPlaybackCodecScreen() {
 		item {
 			ListSection(
 				overlineContent = { Text(stringResource(R.string.pref_playback_advanced).uppercase()) },
-				headingContent = { Text(stringResource(R.string.pref_codec_tweaking)) },
-				captionContent = { Text(stringResource(R.string.codec_level_warning)) },
+				headingContent = { Text(stringResource(R.string.preference_codecs)) },
+				captionContent = { Text(stringResource(R.string.preference_codecs_summary)) },
 			)
 		}
