+++ app/src/main/res/values/strings.xml
@@ -542,22 +542,6 @@
     <string name="rewind">Rewind</string>
     <string name="preference_enable_libass">Enable Advanced SubStation Alpha subtitle client rendering</string>
     <string name="preference_enable_pgs">Direct play PGS subtitles</string>
-    <string name="user_avc_level">H.264 (AVC) Level Override</string>
-    <string name="user_hevc_level">H.265 (HEVC) Level Override</string>
-    <string name="codec_level_warning">Changing from Auto may cause playback issues if your device doesn\'t actually support the selected level.</string>
-    <string name="pref_codec_tweaking">Codec Tweaking</string>
-    <string name="pref_codec_tweaking_summary">Override device-reported codec capabilities</string>
-    <string name="codec_level_manual">Manual Override</string>
-    <string name="codec_level_6_2">Level 6.2</string>
-    <string name="codec_level_6_1">Level 6.1</string>
-    <string name="codec_level_6_0">Level 6.0</string>
-    <string name="codec_level_5_2">Level 5.2</string>
-    <string name="codec_level_5_1">Level 5.1</string>
-    <string name="codec_level_5_0">Level 5.0</string>
-    <string name="codec_level_4_2">Level 4.2</string>
-    <string name="codec_level_4_1">Level 4.1</string>
-    <string name="codec_level_4_0">Level 4.0</string>
-    <string name="auto">Auto</string>
     <string name="speech_error_no_permission">No permission to use the microphone. Please enable it in system settings.</string>
     <string name="speech_error_unavailable">Speech recognition is unavailable on your device.</string>
     <string name="speech_error_unknown">An unexpected error occurred during speech recognition.</string>
@@ -595,6 +579,22 @@
     <string name="filters">Filters</string>
     <string name="backdrop_behavior_with_blur">Backdrop with blur</string>
     <string name="backdrop_behavior_without_blur">Backdrop without blur</string>
+    <string name="user_avc_level">H.264 (AVC) level</string>
+    <string name="user_hevc_level">H.265 (HEVC) level</string>
+    <string name="codec_level_warning">Changing from auto may cause playback issues if your device doesn\'t actually support the selected level.</string>
+    <string name="preference_codecs">Media codecs</string>
+    <string name="preference_codecs_summary">Override device-reported codec capabilities</string>
+    <string name="codec_level_manual">Manual override</string>
+    <string name="codec_level_6_2">Level 6.2</string>
+    <string name="codec_level_6_1">Level 6.1</string>
+    <string name="codec_level_6_0">Level 6.0</string>
+    <string name="codec_level_5_2">Level 5.2</string>
+    <string name="codec_level_5_1">Level 5.1</string>
+    <string name="codec_level_5_0">Level 5.0</string>
+    <string name="codec_level_4_2">Level 4.2</string>
+    <string name="codec_level_4_1">Level 4.1</string>
+    <string name="codec_level_4_0">Level 4.0</string>
+    <string name="auto">Auto</string>
     <plurals name="seconds">
         <item quantity="one">%1$s second</item>
         <item quantity="other">%1$s seconds</item>
